Revert "Create the CompileView attribute"
This reverts commit 2417ecbd8c0a1e0720b97ca96009088feaf525be.
