+++ subprojects/core/src/main/java/org/gradle/api/internal/artifacts/JavaEcosystemSupport.java
@@ -31,7 +31,6 @@ import org.gradle.api.attributes.HasAttributes;
 import org.gradle.api.attributes.LibraryElements;
 import org.gradle.api.attributes.MultipleCandidatesDetails;
 import org.gradle.api.attributes.Usage;
-import org.gradle.api.attributes.CompileView;
 import org.gradle.api.attributes.java.TargetJvmEnvironment;
 import org.gradle.api.attributes.java.TargetJvmVersion;
 import org.gradle.api.internal.ReusableAction;
@@ -53,7 +52,6 @@ public abstract class JavaEcosystemSupport {
     public static void configureSchema(AttributesSchema attributesSchema, final ObjectFactory objectFactory) {
         configureUsage(attributesSchema, objectFactory);
         configureLibraryElements(attributesSchema, objectFactory);
-        configureView(attributesSchema, objectFactory);
         configureBundling(attributesSchema);
         configureTargetPlatform(attributesSchema);
         configureTargetEnvironment(attributesSchema);
@@ -61,7 +59,6 @@ public abstract class JavaEcosystemSupport {
         attributesSchema.attributeDisambiguationPrecedence(
                 Category.CATEGORY_ATTRIBUTE,
                 Usage.USAGE_ATTRIBUTE,
-                CompileView.VIEW_ATTRIBUTE,
                 TargetJvmVersion.TARGET_JVM_VERSION_ATTRIBUTE,
                 LibraryElements.LIBRARY_ELEMENTS_ATTRIBUTE,
                 Bundling.BUNDLING_ATTRIBUTE,
@@ -119,65 +116,6 @@ public abstract class JavaEcosystemSupport {
         libraryElementsSchema.getDisambiguationRules().add(LibraryElementsDisambiguationRules.class, actionConfiguration -> {
             actionConfiguration.params(objectFactory.named(LibraryElements.class, LibraryElements.JAR));
         });
-    }
-
-    private static void configureView(AttributesSchema attributesSchema, final ObjectFactory objectFactory) {
-        AttributeMatchingStrategy<CompileView> viewSchema = attributesSchema.attribute(CompileView.VIEW_ATTRIBUTE);
-        viewSchema.getCompatibilityRules().add(CompileViewCompatibilityRules.class);
-        viewSchema.getDisambiguationRules().add(CompileViewDisambiguationRules.class, actionConfiguration -> {
-            actionConfiguration.params(objectFactory.named(CompileView.class, CompileView.JAVA_API));
-            actionConfiguration.params(objectFactory.named(CompileView.class, CompileView.JAVA_COMPLETE));
-        });
-    }
-
-    @VisibleForTesting
-    public static class CompileViewDisambiguationRules implements AttributeDisambiguationRule<CompileView>, ReusableAction {
-
-        final CompileView javaApi;
-        final CompileView javaComplete;
-
-        @Inject
-        public CompileViewDisambiguationRules(
-                CompileView javaApi,
-                CompileView javaComplete) {
-            this.javaApi = javaApi;
-            this.javaComplete = javaComplete;
-        }
-
-        @Override
-        public void execute(MultipleCandidatesDetails<CompileView> details) {
-            Set<CompileView> candidateValues = details.getCandidateValues();
-            CompileView consumerValue = details.getConsumerValue();
-            if (consumerValue == null) {
-                if (candidateValues.contains(javaApi)) {
-                    // Use the api when nothing has been requested
-                    details.closestMatch(javaApi);
-                }
-            } else if (candidateValues.contains(consumerValue)) {
-                // Use what they requested, if available
-                details.closestMatch(consumerValue);
-            }
-        }
-    }
-
-    @VisibleForTesting
-    public static class CompileViewCompatibilityRules implements AttributeCompatibilityRule<CompileView>, ReusableAction {
-
-        @Override
-        public void execute(CompatibilityCheckDetails<CompileView> details) {
-            CompileView consumerValue = details.getConsumerValue();
-            CompileView producerValue = details.getProducerValue();
-            if (consumerValue == null) {
-                // consumer didn't express any preferences, everything fits
-                details.compatible();
-                return;
-            }
-            // The API view is a subset of the complete view, so the complete view can fulfill
-            // a request for the API.
-            if (CompileView.JAVA_API.equals(consumerValue.getName()) && CompileView.JAVA_COMPLETE.equals(producerValue.getName())) {
-                details.compatible();
-            }
-        }
     }
 
     @VisibleForTesting
@@ -250,20 +188,15 @@ public abstract class JavaEcosystemSupport {
         );
         @Override
         public void execute(CompatibilityCheckDetails<Usage> details) {
-            Usage consumerValue = details.getConsumerValue();
-            Usage producerValue = details.getProducerValue();
-            if (consumerValue == null) {
-                // consumer didn't express any preferences, everything fits
-                details.compatible();
-                return;
-            }
-            if (consumerValue.getName().equals(Usage.JAVA_API)) {
-                if (COMPATIBLE_WITH_JAVA_API.contains(producerValue.getName())) {
+            String consumerValue = details.getConsumerValue().getName();
+            String producerValue = details.getProducerValue().getName();
+            if (consumerValue.equals(Usage.JAVA_API)) {
+                if (COMPATIBLE_WITH_JAVA_API.contains(producerValue)) {
                     details.compatible();
                 }
                 return;
             }
-            if (consumerValue.getName().equals(Usage.JAVA_RUNTIME) && producerValue.getName().equals(DEPRECATED_JAVA_RUNTIME_JARS)) {
+            if (consumerValue.equals(Usage.JAVA_RUNTIME) && producerValue.equals(DEPRECATED_JAVA_RUNTIME_JARS)) {
                 details.compatible();
                 return;
             }
