+++ subprojects/dependency-management/src/integTest/groovy/org/gradle/integtests/resolve/bundling/JavaBundlingResolveIntegrationTest.groovy
@@ -155,7 +155,7 @@ class JavaBundlingResolveIntegrationTest extends AbstractModuleDependencyResolve
 
         then:
         if (shouldFail) {
-            failure.assertHasCause("No matching variant of org:producer:1.0 was found. The consumer was configured to find a component for use during compile-time, and its dependencies repackaged (shadow jar) but:")
+            failure.assertHasCause("No matching variant of org:producer:1.0 was found. The consumer was configured to find an API of a component, and its dependencies repackaged (shadow jar) but:")
         } else {
             resolve.expectGraph {
                 root(":", ":test:") {
