+++ subprojects/plugins/src/integTest/groovy/org/gradle/java/JavaLibraryConsumptionIntegrationTest.groovy
@@ -81,17 +81,17 @@ class JavaLibraryConsumptionIntegrationTest extends AbstractIntegrationSpec {
         fails 'resolve'
 
         then:
-        failure.assertHasCause """The consumer was configured to find a component for use during runtime, compatible with Java 7, and its dependencies declared externally. However we cannot choose between the following variants of org.junit.jupiter:junit-jupiter-api:5.6.0:
+        failure.assertHasCause """The consumer was configured to find a runtime of a component compatible with Java 7, and its dependencies declared externally. However we cannot choose between the following variants of org.junit.jupiter:junit-jupiter-api:5.6.0:
   - javadocElements
   - sourcesElements
 All of them match the consumer attributes:
-  - Variant 'javadocElements' capability org.junit.jupiter:junit-jupiter-api:5.6.0 declares a component for use during runtime, and its dependencies declared externally:
+  - Variant 'javadocElements' capability org.junit.jupiter:junit-jupiter-api:5.6.0 declares a runtime of a component, and its dependencies declared externally:
       - Unmatched attributes:
           - Provides documentation but the consumer didn't ask for it
           - Provides javadocs but the consumer didn't ask for it
           - Doesn't say anything about its target Java version (required compatibility with Java 7)
           - Provides release status but the consumer didn't ask for it
-  - Variant 'sourcesElements' capability org.junit.jupiter:junit-jupiter-api:5.6.0 declares a component for use during runtime, and its dependencies declared externally:
+  - Variant 'sourcesElements' capability org.junit.jupiter:junit-jupiter-api:5.6.0 declares a runtime of a component, and its dependencies declared externally:
       - Unmatched attributes:
           - Provides documentation but the consumer didn't ask for it
           - Provides sources but the consumer didn't ask for it
@@ -99,8 +99,8 @@ All of them match the consumer attributes:
           - Provides release status but the consumer didn't ask for it
 The following variants were also considered but didn't match the requested attributes:
   - Variant 'apiElements' capability org.junit.jupiter:junit-jupiter-api:5.6.0 declares a component, and its dependencies declared externally:
-      - Incompatible because this component declares a component for use during compile-time, compatible with Java 8 and the consumer needed a component for use during runtime, compatible with Java 7
-  - Variant 'runtimeElements' capability org.junit.jupiter:junit-jupiter-api:5.6.0 declares a component for use during runtime, and its dependencies declared externally:
-      - Incompatible because this component declares a component, compatible with Java 8 and the consumer needed a component, compatible with Java 7"""
+      - Incompatible because this component declares an API of a component compatible with Java 8 and the consumer needed a runtime of a component compatible with Java 7
+  - Variant 'runtimeElements' capability org.junit.jupiter:junit-jupiter-api:5.6.0 declares a runtime of a component, and its dependencies declared externally:
+      - Incompatible because this component declares a component compatible with Java 8 and the consumer needed a component compatible with Java 7"""
     }
 }
